pyproject.toml
[tool.poetry]
name = "watercrest"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
# TODO: fix up dependencies. Only pydantic is required in the library
# sympy is a dev dependency, everything else is for testing/formatting/linting
python = "^3.10"
black = "^23.10.0"
isort = "^5.12.0"
ruff = "^0.1.1"
pydantic = "^2.4.2"
pytest = "^7.4.2"
pyright = "^1.1.332"
sympy = "^1.12"
tomli-w = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.mccabe]
max-complexity = 10

[tool.black]
line-length = 120
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 120
include_trailing_comma = true

[tool.pyright]
include = ["watercrest"]
pythonVersion = "3.10"